requirements.txt
setuptools==67.8.0
pywin32==305
altgraph==0.17.3
APScheduler==3.10.1
backports.zoneinfo==0.2.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
cx-Oracle==8.3.0
cycler==0.11.0
distlib==0.3.6
fonttools==4.40.0
greenlet==2.0.2
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1 
numpy==1.24.3
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
pefile==2023.2.7
Pillow==9.5.0
pyodbc==4.0.35
premailer==3.10.0 
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.1 
requests_download==0.1.2
six==1.16.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
XlsxWriter==3.1.2
yarg==0.1.9
zipp==3.15.0
zstandard==0.21.0